// app/enums/list-action-type.enum.ts
export enum ListsActionTypes {
  FETCH_STORE_DETAILS = '[StoreDetails] FETCH_STORE_DETAILS',
  FETCH_STORE_DETAILS_SUCCESS = '[StoreDetails] FETCH_STORE_DETAILS_SUCCESS',
  FETCH_STORE_DETAILS_FAILURE = '[StoreDetails] FETCH_STORE_DETAILS_FAILURE',

  FETCH_HEADER_DETAILS = '[StoreDetails] FETCH_HEADER_DETAILS',
  FETCH_HEADER_DETAILS_SUCCESS = '[StoreDetails] FETCH_HEADER_DETAILS_SUCCESS',
  FETCH_HEADER_DETAILS_FAILURE = '[StoreDetails] FETCH_HEADER_DETAILS_FAILURE',

  FETCH_OPPS_FOR_LIST = '[OpportunitiesForList] FETCH_OPPS_FOR_LIST',
  FETCH_OPPS_FOR_LIST_SUCCESS = '[OpportunitiesForList] FETCH_OPPS_FOR_LIST_SUCCESS',
  FETCH_OPPS_FOR_LIST_FAILURE = '[OpportunitiesForList] FETCH_OPPS_FOR_LIST_FAILURE',

  FETCH_LIST_PERFORMANCE_VOLUME = '[List Performance] FETCH_LIST_PERFORMANCE_VOLUME',
  FETCH_LIST_PERFORMANCE_VOLUME_SUCCESS = '[List Performance] FETCH_LIST_PERFORMANCE_VOLUME_SUCCESS',
  FETCH_LIST_PERFORMANCE_VOLUME_ERROR = '[List Performance] FETCH_LIST_PERFORMANCE_VOLUME_ERROR',

  FETCH_LIST_PERFORMANCE_POD = '[List Performance] FETCH_LIST_PERFORMANCE_POD',
  FETCH_LIST_PERFORMANCE_POD_SUCCESS = '[List Performance] FETCH_LIST_PERFORMANCE_POD_SUCCESS',
  FETCH_LIST_PERFORMANCE_POD_ERROR = '[List Performance] FETCH_LIST_PERFORMANCE_POD_ERROR',

  PATCH_LIST = '[List] PATCH_LIST',
  PATCH_LIST_SUCCESS = '[List] PATCH_LIST_SUCCESS',
  PATCH_LIST_FAILURE = '[List] PATCH_LIST_FAILURE',
}